[TSN-262] Create lock wrapper
Создал LockWrapper - обертку над ts-ными объектами, позволяющими использовать их в c++ коде (реализовано добавлением рута, указывающем на этот объект).
Добавил снятие mark с рутов в ручную т.к. sweep не будет снимать с них "пометки" потому что руты в враппере не попадают в управляемую кучу GC (_heap).
Добавил тесты в gtest и ts.

Reviewed-on: https://gitea.devos.club/antiq/tsnative/pulls/437

+++ std/src/tsobject.cpp
@@ -24,8 +24,6 @@
 #include "std/private/allocator.h"
 #include "std/runtime.h"
 
-#include "std/private/logger.h"
-
 static String* superKey = new String("super");
 static String* parentKey = new String("parent");
 
@@ -375,7 +373,7 @@ void* Object::operator new(std::size_t n)
     }
 
     auto* allocator = Runtime::getAllocator();
-    return allocator->allocateObject(static_cast<double>(n));
+    return allocator->allocateObject(n);
 }
 
 class String;
